refactor!: make `simapp.MakeTestEncodingConfig` private (#12747)
* feat: make `simapp.MakeTestEncodingConfig` private

* fix legacy build

* update changelog and upgrading.md

* updates docs
